+++ src/back/InSeconds.Api/Features/Sessions/SubmitAnswer/Response.cs
@@ -6,6 +6,6 @@ public sealed record SubmitAnswerResponse(
     int Score,
     string CorrectArtist,
     string CorrectTitle,
-    int ListenedDurationSeconds,
+    decimal ListenedDurationSeconds,
     double? AverageSecondsWhenCorrect,
     double FailureRatePercent);
